Исправление ошибок найденных при обновлении ядра проекта Ringoman до версии 2.6 [banners]  - Исправлен выбор и обработка дат, в шаблоне редактирования баннера  - Добавлены тексты ошибок  - Обновлены шаблоны модуля  - Исправлена работа с таблицами в классе АПИ модуля  - Исправлены ошибки в страницах администрирования
[catsubcat]
 - Убрана корзина из меню модуля
 - Исправлен установщик модуля
 - Исправлена ошибка в шаблоне редактирования страницы
 - Исправлена генерация текстового идентификатора в методе CCategorySubCategory::_BeforeSave
 - Исправлено удаление текстовых страниц
 - Обновлены вызовы в main.inc.php
 - Исправлены ошибки и стиль кода в административной странице
 - Исправлена страница настройки модуля
 - Обновлена генерация дерева модуля
 - Обновлён шаблона анонсов записей

[fm]
 - Исправлен инсталятор модуля

[guestbook2]
 - Исправлены вызовы определения таблиц

[hints]
 - Добавлено подтверждение удаления подсказки
 - Добавлено автозаполнение кодов подсказок по шаблонам
 - Исправлены ошибки в админке

[interfaces]
 - Исправлена обработка даты в фильтре
 - Исправлено определение браузера в виджете BrowserAttention

[main]
 - Добавлены новые тексты ошибок
 - Исправлен класс CModulesAccess.php
 - Исправлено определение пустоты id при сохранении записи _SaveData
 - Исправлена обработка полей float|double
 - Расширен функционал обработчика дерева сайта
 - Исправлена обработка массивов в CUrlParser
 - Исправлена деавторизация в ПЧ сайта
 - Исправлены ошибки в редактировании учётных записей пользователей
 - Исправлены ошибки при редактировании настроек модуля
 - Исправлены ошибки в виджете Pic

[navigation]
 - Добавлен экспорт типов меню
 - Оптимизирован поиск пунктов меню
 - Исправлены ошибки в административном интерфейсе

[wave]
 - Обновлён код класса АПИ

+++ modules/hints/install/templates/admin/hints.tpl
@@ -50,7 +50,7 @@
 					<a href="{get_url action=edit id=$oItem.id}" title="{#edit#}">
 						<img src="{#images_path#}/icons2/edit.gif" alt="{#edit#}" />
 					</a>
-					<a href="{get_url action=delete id=$oItem.id}" title="{#delete#}">
+					<a href="{get_url action=delete id=$oItem.id}" title="{#delete#}" onclick="return confirm('{#delete_confirm#}');">
 						<img src="{#images_path#}/icons2/delete.gif" alt="{#delete#}" />
 					</a>
 				</div>
